// src/common/types/withModel.ts
interface IWithModelBodyParams {
    pageIndex: number; // required, номер страницы
    pageSize: number; // required, количество элементов на странице
    sortField?: string; // optional, поле для сортировки
    sortDirection?: string; // optional, ENUM: ['ASC', 'DESC'], порядок сортировки
    searchBy: {
        code?: string; //  optional, поиск по ЛМ коду товара (логическое или }
        adeoCode?: string; // optional, поиск по коду ADEO
        description?: string; // optional, название товара
        ean?: string; // optionaд, ШК товара
        supplierCode?: string; // optional
        supplierName?: string; // optional
        supplierTaxIdentifier?: string; // optional
        productManagementNomenclature?: {
            departmentId?: number; // optional, департамент отдела товара
            subDepartmentId?: number; // optional, поддепартамент отдела товара
            typeId?: number; // optional, категория товара
            subTypeId?: number; // optional, подкатегория товара
        }[];
        qualityModelId?: number; // optional, номер модели качества связанной с товаром
        productModelNomenclature?: {
            //- номенклатура товарной модели
            // modelDepartmentId?: string; // optional- код отдела
            // modelSubDepartmentId?: string; // optional - код подотдела
            // modelConsolidationId?: string; // optional - код группы
            // modelCodeId?: string; // optional - код модели

            productModelNomenclatureModelCode?: string;
            productModelNomenclatureConsolidationCode?: string;
            productModelNomenclatureSubDepartmentCode?: string;
            productModelNomenclatureDepartmentCode?: string;
        }[];
        range?: string; // optional
        buCode?: number[]; // optional
        withQualityModel?: boolean; // optoional, поиск по наличию qualityModelId в теле сущности
        mdd?: boolean; // optional, поиск по товарам СТМ
        import?: boolean; //optional, поиск по импорту
        attributes?: {
            // optional
            attributeCode: string; // required
            values?: string[]; // optional
        };
        project?: boolean; // optional
        blockType?: string; // optional
        productWithSubstances?: string; // optional
        regulatoryStatuses?: string[]; // (логическое или) ??
        conformityStatuses?: string[]; // (логическое или) ??
    };
}

interface IWithModelHeaderParams {
    securityCode: string;
}

export interface IWithModelParams {
    header: IWithModelHeaderParams;
    body: IWithModelBodyParams;
}